Changed the size argument to be a size_t type, to make the prototype look like the system one as otherwise the compiler complains a lot.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@805 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/common/sprintf.h
@@ -20,6 +20,6 @@
 #ifndef __SPRINTF_H__
 #define __SPRINTF_H__
 
-int snprintf (char *buf, int size, const char *fmt, ...);
+int snprintf (char *buf, size_t size, const char *fmt, ...);
 
 #endif /* __SPRINTF_H__ */
